2020-2021/StudentProjects/AlwaysRight-SocialDistancing/yolo-server/requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.3
Pillow==8.0.1
requests==2.21.0
six==1.12.0
urllib3==1.24.3
Werkzeug==0.15.2
yolo34py==0.1rc12
